pyproject.toml
[tool.poetry]
name = "api-gerenciamento-consultas-medicas"
version = "0.1.0"
description = "API RESTful de Gerenciamento de Consultas Médicas"
authors = ["Miguel Lucas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
django-cors-headers = "^4.3"
django-filter = "^24.3"
drf-spectacular = "^0.27"
psycopg2-binary = "^2.9"
python-decouple = "^3.8"
gunicorn = "^23.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4"
black = "^24.4"
pytest = "^8.2"
pytest-django = "^4.8"
factory-boy = "^3.3"
coverage = "^7.5"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
